feature wrap config and client for #525
