+++ kube-derive/src/custom_resource.rs
@@ -164,8 +164,8 @@ pub(crate) fn derive(input: proc_macro2::TokenStream) -> proc_macro2::TokenStrea
         impl #rootident {
             pub fn new(name: &str, spec: #ident) -> Self {
                 Self {
-                    api_version: <#rootident as kube_core::Resource>::api_version(&()).to_string(),
-                    kind: <#rootident as kube_core::Resource>::kind(&()).to_string(),
+                    api_version: <#rootident as kube::Resource>::api_version(&()).to_string(),
+                    kind: <#rootident as kube::Resource>::kind(&()).to_string(),
                     metadata: k8s_openapi::apimachinery::pkg::apis::meta::v1::ObjectMeta {
                         name: Some(name.to_string()),
                         ..Default::default()
@@ -184,7 +184,7 @@ pub(crate) fn derive(input: proc_macro2::TokenStream) -> proc_macro2::TokenStrea
 
     let api_ver = format!("{}/{}", group, version);
     let impl_resource = quote! {
-        impl kube_core::Resource for #rootident {
+        impl kube::Resource for #rootident {
             type DynamicType = ();
 
             fn group(_: &()) -> std::borrow::Cow<'_, str> {
@@ -223,8 +223,8 @@ pub(crate) fn derive(input: proc_macro2::TokenStream) -> proc_macro2::TokenStrea
             impl Default for #rootident {
                 fn default() -> Self {
                     Self {
-                        api_version: <#rootident as kube_core::Resource>::api_version(&()).to_string(),
-                        kind: <#rootident as kube_core::Resource>::kind(&()).to_string(),
+                        api_version: <#rootident as kube::Resource>::api_version(&()).to_string(),
+                        kind: <#rootident as kube::Resource>::kind(&()).to_string(),
                         metadata: k8s_openapi::apimachinery::pkg::apis::meta::v1::ObjectMeta::default(),
                         spec: Default::default(),
                         #statusdef
@@ -358,8 +358,8 @@ pub(crate) fn derive(input: proc_macro2::TokenStream) -> proc_macro2::TokenStrea
                     .expect("valid custom resource from #[kube(attrs..)]")
             }
 
-            pub fn api_resource() -> kube_core::api_resource::ApiResource {
-                kube_core::api_resource::ApiResource::erase::<Self>(&())
+            pub fn api_resource() -> kube::core::ApiResource {
+                kube::core::ApiResource::erase::<Self>(&())
             }
         }
     };
