+++ kube/Cargo.toml
@@ -20,6 +20,8 @@ default = ["native-tls"]
 native-tls = ["openssl", "hyper-tls", "tokio-native-tls"]
 rustls-tls = ["hyper-rustls", "tokio-rustls"]
 derive = ["kube-derive"]
+client = ["hyper", "tower", "hyper-timeout", "pin-project", "chrono", "jsonpath_lib"]
+config = ["pem", "dirs"]
 jsonpatch = ["json-patch"]
 ws = ["tokio-tungstenite", "rand", "kube-core/ws"]
 oauth = ["tame-oauth"]
@@ -27,14 +29,14 @@ gzip = ["async-compression"]
 admission = ["json-patch", "kube-core/admission"]
 
 [package.metadata.docs.rs]
-features = ["derive", "ws", "oauth", "jsonpatch"]
+features = ["derive", "client", "config", "ws", "oauth", "jsonpatch"]
 # Define the configuration attribute `docsrs`. Used to enable `doc_cfg` feature.
 rustdoc-args = ["--cfg", "docsrs"]
 
 [dependencies]
 base64 = "0.13.0"
-chrono = "0.4.19"
-dirs = { package = "dirs-next", version = "2.0.0" }
+chrono = { version = "0.4.19", optional = true }
+dirs = { package = "dirs-next", optional = true, version = "2.0.0" }
 serde = { version = "1.0.118", features = ["derive"] }
 serde_json = "1.0.61"
 serde_yaml = "0.8.17"
@@ -44,7 +46,7 @@ log = "0.4.11"
 either = "1.6.1"
 thiserror = "1.0.23"
 futures = "0.3.8"
-pem = "0.8.2"
+pem = { version = "0.8.2", optional = true }
 openssl = { version = "0.10.32", optional = true }
 tokio-native-tls = { version = "0.3.0", optional = true }
 tokio-rustls = { version = "0.22.0", features = ["dangerous_configuration"], optional = true }
@@ -53,18 +55,18 @@ tokio = { version = "1.0.1", features = ["time", "signal", "sync"] }
 static_assertions = "1.1.0"
 kube-derive = { path = "../kube-derive", version = "^0.54.0", optional = true }
 kube-core = { path = "../kube-core", version = "^0.54.0"}
-jsonpath_lib = "0.2.6"
+jsonpath_lib = { version = "0.2.6", optional = true }
 tokio-util = { version = "0.6.0", features = ["io", "codec"] }
 json-patch = { version = "0.2.6", optional = true }
-hyper = { version = "0.14.2", features = ["client", "http1", "stream", "tcp"] }
+hyper = { version = "0.14.2", optional = true, features = ["client", "http1", "stream", "tcp"] }
 hyper-tls = { version = "0.5.0", optional = true }
 hyper-rustls = { version = "0.22.1", optional = true }
 tokio-tungstenite = { version = "0.14.0", optional = true }
-tower = { version = "0.4.6", features = ["buffer", "util"] }
+tower = { version = "0.4.6", optional = true, features = ["buffer", "util"] }
 async-compression = { version = "0.3.7", features = ["gzip", "tokio"], optional = true }
-hyper-timeout = "0.4.1"
+hyper-timeout = {version = "0.4.1", optional = true }
 tame-oauth = { version = "0.4.7", features = ["gcp"], optional = true }
-pin-project = "1.0.4"
+pin-project = { version = "1.0.4", optional = true }
 rand = { version = "0.8.3", optional = true }
 tracing = "0.1.25"
